[Refactor] Axios 인스턴스 생성 로직 분리
- client.ts: 클라이언트 측에서 사용할 axios 인스턴스
- server.ts: 서버 측에서 사용할 axios 인스턴스
