+++ src/api/client.ts
@@ -1,7 +1,7 @@
 import axios from 'axios';
 import { AxiosInstance, AxiosError, InternalAxiosRequestConfig, AxiosResponse } from 'axios';
 
-const Instance = (): AxiosInstance => {
+const getMoneedInstance = (): AxiosInstance => {
     const instance: AxiosInstance = axios.create({
         withCredentials: true,
         baseURL: process.env.NEXT_PUBLIC_MONEED_BASE_URL,
@@ -37,4 +37,4 @@ const Instance = (): AxiosInstance => {
     return instance;
 };
 
-export const http = Instance();
+export const http = getMoneedInstance();
