+++ src/api/kakao.api.ts
@@ -1,38 +1,12 @@
 import 'server-only';
-import { KakaoRefreshTokenResponse, KakaoTokenResponse } from '@/types/kakao';
+import { KakaoRefreshTokenResponse, KakaoTokenResponse, KakaoUserInfo } from '@/types/kakao';
 import axios, { AxiosError } from 'axios';
-import { ProviderRepository } from '@/repositories/provider.repository';
-import { deleteSession } from '@/lib/session';
-
-export interface KakaoUserInfo {
-    id: bigint;
-    connected_at: string;
-    synched_at: string;
-    properties: {
-        nickname: string;
-        profile_image: string;
-        thumbnail_image: string;
-    };
-    kakao_account: {
-        profile: {
-            nickname: string;
-            thumbnail_image_url: string;
-            profile_image_url: string;
-        };
-        age_range: string;
-        gender: 'male' | 'female';
-        name: string;
-        email: string;
-        birthyear: string;
-        birthday: string;
-    };
-}
+import { kakao } from './server';
 
 const kakaoTokenUrl = 'https://kauth.kakao.com/oauth/token';
 const kakaoUserInfoUrl = 'https://kapi.kakao.com/v2/user/me';
 const kakaoLogoutUrl = 'https://kapi.kakao.com/v1/user/logout';
 const kakaoLeaveUrl = 'https://kapi.kakao.com/v1/user/unlink';
-const providerRepository = new ProviderRepository();
 
 export const getKakaoToken = async (code: string) => {
     try {
@@ -76,69 +50,9 @@ export const refreshKakaoToken = async (refreshToken: string) => {
     return res.data;
 };
 
-const kakaoAuthInstance = (() => {
-    const instance = axios.create();
-
-    instance.interceptors.request.use(async config => {
-        try {
-            if (config.headers.Authorization) {
-                const token = (config.headers.Authorization as string).split(' ')[1];
-                const providerData = await providerRepository.getTokenExpiration('kakao', token);
-                if (providerData) {
-                    const isAcessTokenExpired = providerData.accessTokenExpiresIn < new Date();
-                    const isRefreshTokenExpired = providerData.refreshTokenExpiresIn < new Date();
-
-                    if (!isAcessTokenExpired && !isRefreshTokenExpired) {
-                        config.headers.Authorization = `Bearer ${providerData.accessToken}`;
-                        return config;
-                    } else if (isAcessTokenExpired) {
-                        const newToken = await refreshKakaoToken(providerData.refreshToken);
-                        config.headers.Authorization = `Bearer ${newToken.access_token}`;
-                        await providerRepository.updateTokenData(
-                            {
-                                provider: 'kakao',
-                                providerUserId: providerData.providerUserId,
-                            },
-                            {
-                                accessToken: newToken.access_token,
-                                refreshToken: newToken.refresh_token,
-                                accessTokenExpiresIn: new Date(Date.now() + newToken.expires_in * 1000),
-                                refreshTokenExpiresIn: newToken.refresh_token_expires_in
-                                    ? new Date(Date.now() + newToken.refresh_token_expires_in * 1000)
-                                    : undefined,
-                            },
-                        );
-                        return config;
-                    } else if (isRefreshTokenExpired) {
-                        await providerRepository.delete('kakao', providerData.providerUserId);
-                        await deleteSession();
-                        return Promise.reject(new AxiosError('Token expired'));
-                    }
-                }
-                return config;
-            }
-            return Promise.reject(new AxiosError('Authorization header is required'));
-        } catch (error) {
-            console.error('Failed to get Kakao token expiration:', (error as AxiosError).response?.data);
-            return Promise.reject(error);
-        }
-    });
-
-    instance.interceptors.response.use(response => {
-        try {
-            return response;
-        } catch (error) {
-            console.error('Failed to get Kakao token expiration:', (error as AxiosError).response?.data);
-            return Promise.reject(error);
-        }
-    });
-
-    return instance;
-})();
-
 export const getKakaoUserInfo = async (accessToken: string) => {
     try {
-        const res = await kakaoAuthInstance.get<KakaoUserInfo>(kakaoUserInfoUrl, {
+        const res = await kakao.get<KakaoUserInfo>(kakaoUserInfoUrl, {
             headers: {
                 'Content-Type': 'application/x-www-form-urlencoded;charset=utf-8',
                 Authorization: `Bearer ${accessToken}`,
@@ -153,7 +67,7 @@ export const getKakaoUserInfo = async (accessToken: string) => {
 
 export const logoutKakao = async ({ accessToken, providerUserId }: { accessToken: string; providerUserId: string }) => {
     try {
-        const res = await kakaoAuthInstance.post(
+        const res = await kakao.post(
             kakaoLogoutUrl,
             {
                 target_id_type: 'user_id',
@@ -175,7 +89,7 @@ export const logoutKakao = async ({ accessToken, providerUserId }: { accessToken
 
 export const leaveKakao = async ({ accessToken, providerUserId }: { accessToken: string; providerUserId: string }) => {
     try {
-        const res = await kakaoAuthInstance.post<{ id: bigint }>(
+        const res = await kakao.post<{ id: bigint }>(
             kakaoLeaveUrl,
             {
                 target_id_type: 'user_id',
