+++ src/api/server.ts
@@ -1,0 +1,100 @@
+import axios from 'axios';
+import 'server-only';
+import { ProviderRepository } from '@/repositories/provider.repository';
+import { deleteSession } from '@/lib/session';
+import { refreshKakaoToken } from './kakao.api';
+import { AxiosError } from 'axios';
+
+const providerRepository = new ProviderRepository();
+
+const getKisInstance = () => {
+    const instance = axios.create({
+        baseURL: process.env.KIS_BASE_URL,
+        headers: {
+            'Content-type': 'application/json; charset=utf-8',
+            appkey: process.env.KIS_APP_KEY,
+            appsecret: process.env.KIS_APP_SECRET,
+        },
+    });
+
+    instance.interceptors.request.use(async config => {
+        // TODO: DB에서 한국투자증권 액세스 토큰 조회
+        // const accessToken = await getAccessToken();
+        // config.headers.Authorization = `Bearer ${accessToken}`;
+        return config;
+    });
+
+    instance.interceptors.response.use(
+        async response => {
+            return response;
+        },
+        async error => {
+            console.error(error.response?.data);
+            return Promise.reject(error);
+        },
+    );
+    return instance;
+};
+
+const getKakaoInstance = () => {
+    const instance = axios.create();
+
+    instance.interceptors.request.use(async config => {
+        try {
+            if (config.headers.Authorization) {
+                const token = (config.headers.Authorization as string).split(' ')[1];
+                const providerData = await providerRepository.getTokenExpiration('kakao', token);
+                if (providerData) {
+                    const isAcessTokenExpired = providerData.accessTokenExpiresIn < new Date();
+                    const isRefreshTokenExpired = providerData.refreshTokenExpiresIn < new Date();
+
+                    if (!isAcessTokenExpired && !isRefreshTokenExpired) {
+                        config.headers.Authorization = `Bearer ${providerData.accessToken}`;
+                        return config;
+                    } else if (isAcessTokenExpired) {
+                        const newToken = await refreshKakaoToken(providerData.refreshToken);
+                        config.headers.Authorization = `Bearer ${newToken.access_token}`;
+                        await providerRepository.updateTokenData(
+                            {
+                                provider: 'kakao',
+                                providerUserId: providerData.providerUserId,
+                            },
+                            {
+                                accessToken: newToken.access_token,
+                                refreshToken: newToken.refresh_token,
+                                accessTokenExpiresIn: new Date(Date.now() + newToken.expires_in * 1000),
+                                refreshTokenExpiresIn: newToken.refresh_token_expires_in
+                                    ? new Date(Date.now() + newToken.refresh_token_expires_in * 1000)
+                                    : undefined,
+                            },
+                        );
+                        return config;
+                    } else if (isRefreshTokenExpired) {
+                        await providerRepository.delete('kakao', providerData.providerUserId);
+                        await deleteSession();
+                        return Promise.reject(new AxiosError('Token expired'));
+                    }
+                }
+                return config;
+            }
+            return Promise.reject(new AxiosError('Authorization header is required'));
+        } catch (error) {
+            console.error('Failed to get Kakao token expiration:', (error as AxiosError).response?.data);
+            return Promise.reject(error);
+        }
+    });
+
+    instance.interceptors.response.use(response => {
+        try {
+            return response;
+        } catch (error) {
+            console.error('Failed to get Kakao token expiration:', (error as AxiosError).response?.data);
+            return Promise.reject(error);
+        }
+    });
+
+    return instance;
+};
+
+export const kis = getKisInstance();
+export const kakao = getKakaoInstance();
